[tmux] Better mouse drag handling in copy mode
If drag ends and the buffer is NOT scrolled up (scroll_position is 0),
copy the selection AND (new) exit the copy mode. This works better for
the usual cases when one drags the mouse to just copy the text from the
visible screen and expects to not enter the copy mode but wants to stay
in the normal mode.